sprint5: modify reports

+++ src/main/java/com/example/onlineshopproject/repository/customs/ProductCustomRepository.java
@@ -3,9 +3,10 @@ package com.example.onlineshopproject.repository.customs;
 import com.example.onlineshopproject.entity.CategoryEntity;
 import com.example.onlineshopproject.entity.ProductEntity;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductCustomRepository {
-    List<ProductEntity> findProductByFilter(CategoryEntity categoryEntity, Double minPrice, Double maxPrice,
+    List<ProductEntity> findProductByFilter(CategoryEntity categoryEntity, BigDecimal minPrice, BigDecimal maxPrice,
                                             Boolean isDiscount, String sort);
 }
